% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{boletin_cve}
\alias{boletin_cve}
\title{Boletin's code}
\usage{
boletin_cve(legislatura, sesion)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}

\item{sesion}{A numeric value above 0.}
}
\value{
Summary of the session plus information about the documents discussed.
}
\description{
Creates the code of the summary of a session of the senate with the
documents that got talk.
}
\examples{
boletin_cve(14, 1)
}
\seealso{
Other generators of CVE: 
\code{\link{document_cve}()},
\code{\link{sumario_cve}()}
}
\concept{generators of CVE}
