% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{check_code}
\alias{check_code}
\title{Check code}
\usage{
check_code(id)
}
\arguments{
\item{id}{Code of an official publication of the Senate}
}
\value{
TRUE if passes all the checks
}
\description{
Check code
}
\examples{
check_code("BOCG_S_1_1")
check_code("BOCG_B_1_1")
check_code("BOCG_D_1_1_1")
}
