% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comisiones_ponencias.R
\name{comisiones}
\alias{comisiones}
\title{\emph{Comisiones}}
\usage{
comisiones(legislatura)
}
\arguments{
\item{legislatura}{A numeric value above 12 (No information prior to then).}
}
\value{
A \code{data.frame} with the \emph{comisiones} and \emph{ponencias} of that legislature.
}
\description{
Look up the \emph{comisiones} and \emph{ponencias} of a legislature.
Rellevant to know the code of the work groups.
}
\examples{
if (interactive()) {
   comisiones(13)
}
}
