% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plenarias.R
\name{detalles}
\alias{detalles}
\title{Check individual session}
\usage{
detalles(url)
}
\arguments{
\item{url}{A url path to a session.}
}
\value{
A \link{data.frame} with all the information about a session.
}
\description{
Check individual session
}
\examples{
if (interactive()) {
    pl <- plenarias(10)
    detalles(pl$fichUrlDetalleSesion[1])
}
}
\seealso{
\code{\link[=plenarias]{plenarias()}}
}
