% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{document_cve}
\alias{document_cve}
\title{Document CVE}
\usage{
document_cve(legislatura, sesion, number)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}

\item{sesion}{A numeric value above 0.}

\item{number}{A numeric value of}
}
\value{
A character vector of a valid CVE.
}
\description{
Code needd to retrieve data about a document.
}
\examples{
document_cve(14, 1, 1)
}
\seealso{
Other generators of CVE: 
\code{\link{boletin_cve}()},
\code{\link{sumario_cve}()}
}
\concept{generators of CVE}
