% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queries.R
\name{get_xml}
\alias{get_xml}
\title{Retrieve the XML content}
\usage{
get_xml(query, encoding)
}
\arguments{
\item{query}{A character with the url of the query to the website.}

\item{encoding}{A character with the encoding, usually either "windows-1252", "ISO-8859-15" or "UTF-8".}
}
\description{
Look up on the internet and get the content
}
\keyword{internal}
