% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composicion.R
\name{grupos}
\alias{grupos}
\title{Political groups and parties}
\usage{
grupos(legislatura)
}
\arguments{
\item{legislatura}{A numeric value above 12 (No information prior to then).}
}
\value{
A matrix with all the information available.
}
\description{
Retrieves the data for each term of the groups and political parties present
}
\examples{
grupos()
}
