% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iniciativas.R
\name{iniciativas}
\alias{iniciativas}
\title{Look up the legislative iniciatives}
\usage{
iniciativas(legislatura)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}
}
\value{
A data.frame with the initiatives of that legislature.
}
\description{
Look up the legislative iniciatives
}
\examples{
if (interactive()) {
    iniciativas(14)
}
}
\seealso{
iniciativas_parlamentarias
}
