% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leyes.R
\name{leyes}
\alias{leyes}
\title{Approved laws}
\usage{
leyes(legislatura)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}
}
\value{
A data.frame with the approved laws and related information.
}
\description{
Check how many laws were approved.
Note it might now return all the laws approved.
}
\examples{
if (interactive()) {
    l <- leyes(14)
    head(l)
}
}
