% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mociones.R
\name{mociones}
\alias{mociones}
\title{\emph{Mociones}}
\usage{
mociones(legislatura)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}
}
\value{
A data.frame with information about the \emph{mociones}.
}
\description{
Download \emph{mociones}
}
\examples{
if (interactive() {
    mociones(13)
}
}
\seealso{
iniciativas_parlamentarias()
}
