% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organigrama.R
\name{organigrama}
\alias{organigrama}
\title{Organization chart}
\usage{
organigrama()
}
\value{
A \code{data.frame} with all the information available.
}
\description{
Retrieves the relationships between people responsible
}
\examples{
head(organigrama())
}
