% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plenarias.R
\name{plenarias}
\alias{plenarias}
\title{Plenary sessions}
\usage{
plenarias(legislatura = 10)
}
\arguments{
\item{legislatura}{A numeric value above 12 (No information prior to then).}
}
\value{
A data.frame with the information available.
}
\description{
How many session have been.
}
\examples{
if (interactive()) {
    head(plenarias(10))
}
}
