% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/senador.R
\name{senador}
\alias{senador}
\title{Retrieves the open data of a senator.}
\usage{
senador(legislatura, codigo)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}

\item{codigo}{Number or idweb of the senator.}
}
\description{
Retrieves the open data of a senator.
}
\examples{
if (interactive()) {
    s <- senador(14, 19050)
    head(s)
}
}
