% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composicion.R
\name{senadores}
\alias{senadores}
\title{Senators since 1977}
\usage{
senadores(legislatura = 12:14)
}
\arguments{
\item{legislatura}{A numeric value above 12 (No information prior to then).}
}
\value{
A \code{data.frame} with the information available.
}
\description{
Past and current appointed members.
}
\examples{
if (interactive()) {
    head(senadores(12))
}
}
