% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{sumario}
\alias{sumario}
\title{Retrieve the document of sumario}
\usage{
sumario(id)
}
\arguments{
\item{id}{The character of the CVE of the document.}
}
\value{
A matrix with a summary of the session, date, documents, authors.
}
\description{
Returns the summary of a session in a tidy way.
If multiple autors are involved in a disposicion and multiple documents they
get mixed.
}
\examples{
sumario_id <- sumario_nbo(14, 1)
s <- sumario(sumario_id)
}
