% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{sumario_cve}
\alias{sumario_cve}
\title{Summary code for a session}
\usage{
sumario_cve(legislatura, sesion)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}

\item{sesion}{A numeric value above 0.}
}
\value{
A character id of the code for the summary of that session.
}
\description{
Creates the code of the summary of a session of the senate.
}
\examples{
sumario_cve(14, 1)
}
\seealso{
Other generators of CVE: 
\code{\link{boletin_cve}()},
\code{\link{document_cve}()}
}
\concept{generators of CVE}
