% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tidy_diputado}
\alias{tidy_diputado}
\title{Clean the data of the diputado}
\usage{
tidy_diputado(diputado, legislatura)
}
\arguments{
\item{diputado}{A numeric id.}

\item{legislatura}{A numeric value of the legislature}
}
\value{
A vector with the information available.
}
\description{
Clean the data of the diputado
}
\examples{
a <- url_diputado(10, 13)
tidy_diputado(a)
}
