% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votos.R
\name{votos}
\alias{votos}
\title{Votos}
\usage{
votos(legislatura, sesion, votacion = NULL)
}
\arguments{
\item{legislatura}{A numeric value. Constituent was 0.}

\item{sesion}{The number of the meeting.}

\item{votacion}{Number of the vote.}
}
\value{
A votos class with information about who voted what.
}
\description{
Retrieve votes in a sesion.
}
\examples{
votos_64 <- votos(14, 64)
votos_3_8 <- votos(14, 3, 8)
}
