% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plenarias.R
\name{detalles}
\alias{detalles}
\title{Check individual session}
\usage{
detalles(url)
}
\arguments{
\item{url}{A url path to a session.}
}
\value{
A \link{data.frame} with all the information about a session.
}
\description{
Check individual session
}
\examples{
url <-  "https://www.senado.es/web/ficopendataservlet?tipoFich=11&legis=14&org=S000040&numSes=020&numConv=01&fecha=17112020"
detalles(url)
}
\seealso{
\code{\link[=plenarias]{plenarias()}}
}
