% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{documento}
\alias{documento}
\title{A document}
\usage{
documento(cve)
}
\arguments{
\item{cve}{A character with a document CVE.}
}
\description{
A document
}
\examples{
document_csv <- "BOCG_D_14_110_901"
if (interactive()) {
    documento(document_csv)
}
}
\seealso{
\code{\link[=document_csv]{document_csv()}}

Other types of documents: 
\code{\link{sumario}()}
}
\concept{types of documents}
