% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organigrama.R
\name{organigrama}
\alias{organigrama}
\title{Organization chart}
\usage{
organigrama(legislatura)
}
\arguments{
\item{legislatura}{A numeric value above 12 (No information prior to then).}
}
\value{
A \code{data.frame} with all the information available.
}
\description{
Retrieves the relationships between people responsible of the Senate.
At the moment due to limitation on the website only works on the 13 legislature.
}
\examples{
head(organigrama(13))
}
