% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{sumario}
\alias{sumario}
\title{Retrieve the document of sumario}
\usage{
sumario(cve)
}
\arguments{
\item{cve}{The character of the CVE of the document.}
}
\value{
A data.frame with a summary of the session, date, documents, authors.
}
\description{
Returns the summary of a session in a tidy way.
If multiple autors are involved in a disposicion and multiple documents they
get mixed.
}
\examples{
sumario_cve <- sumario_nbo(14, 1)
head(sumario(sumario_cve))
}
\seealso{
Other types of documents: 
\code{\link{documento}()}
}
\concept{types of documents}
