% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diario.R
\name{sumario}
\alias{sumario}
\title{Retrieve the document of \emph{sumario}}
\usage{
sumario(cve)
}
\arguments{
\item{cve}{The character of the CVE of the document.}
}
\value{
A data.frame with a summary of the session, date, documents, authors.
}
\description{
Returns the summary of a session in a tidy way.
If multiple authors are involved in a disposition and multiple documents they
get mixed.
}
\examples{
sumario_csv <- sumario_csv(14, 3)
if (interactive()) {
    head(sumario(sumario_csv))
}
}
\seealso{
Other types of documents: 
\code{\link{documento}()}
}
\concept{types of documents}
