% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{tidy_diputado}
\alias{tidy_diputado}
\title{Clean the data of the diputado on the congress!!}
\usage{
tidy_diputado(legislatura, diputado)
}
\arguments{
\item{legislatura}{A numeric value of the legislature}

\item{diputado}{A numeric id.}
}
\value{
A vector with the information available.
}
\description{
Clean the data of the diputado on the congress!!
}
\examples{
tidy_diputado(10, 13)
}
